+++ docs/examples/controlpanels/get_types.ts
@@ -1,0 +1,6 @@
+import { Client } from '@robgietema/nick';
+
+const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
+const login = await cli.login({ username: 'admin', password: 'admin' });
+
+const { data } = await cli.getControlpanel({ path: '/dexterity-types' });